[services] Blob - Fix time types
Summary:
We want to use fixed-length variables like `uint64_t` instead of `long long` which size might be compiler-dependent.

https://linear.app/comm/issue/ENG-584/refactor-time-types-in-c

Test Plan:
```
cd services
yarn test-blob-service
yarn test-blob-service-dev-mode
```
Works as it did before

Reviewers: palys-swm, geekbrother, ashoat

Reviewed By: palys-swm, ashoat

Subscribers: benschac, ashoat, palys-swm, Adrian, atul

Differential Revision: https://phabricator.ashoat.com/D2952

+++ services/blob/docker-server/contents/server/src/DatabaseEntities/BlobItem.h
@@ -13,7 +13,7 @@ class BlobItem : public Item {
 
   std::string blobHash;
   S3Path s3Path;
-  long long created = 0;
+  uint64_t created = 0;
 
   void validate() const override;
 
@@ -28,7 +28,7 @@ public:
   BlobItem(
       const std::string blobHash,
       const S3Path s3Path,
-      long long created = 0);
+      uint64_t created = 0);
   BlobItem(const AttributeValues &itemFromDB);
 
   void assignItemFromDatabase(const AttributeValues &itemFromDB) override;
@@ -38,7 +38,7 @@ public:
 
   std::string getBlobHash() const;
   S3Path getS3Path() const;
-  long long getCreated() const;
+  uint64_t getCreated() const;
 };
 
 } // namespace database
